registers typeorm manually, restructure datasources
