+++ src/app.module.ts
@@ -54,7 +54,7 @@ import { TargetedMessagingModule } from '@/routes/targeted-messaging/targeted-me
 import { PostgresDatabaseModule } from '@/datasources/db/postgres-database.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import { postgresFactory } from '@/config/entities';
+import { postgresConfig } from '@/config/entities/postgres.config';
 
 @Module({})
 export class AppModule implements NestModule {
@@ -133,8 +133,12 @@ export class AppModule implements NestModule {
         }),
         TypeOrmModule.forRootAsync({
           imports: [ConfigModule],
-          useFactory: (configService: ConfigService) =>
-            postgresFactory(configService),
+          useFactory: (configService: ConfigService) => {
+            return {
+              ...{ autoLoadEntities: true, manualInitialization: true },
+              ...postgresConfig(configService.getOrThrow('db.postgres')),
+            };
+          },
           inject: [ConfigService],
         }),
       ],
