+++ src/config/entities/__tests__/configuration.ts
@@ -86,7 +86,7 @@ export default (): ReturnType<typeof configuration> => ({
   },
   db: {
     postgres: {
-      schema: faker.string.sample(),
+      schema: process.env.POSTGRES_SCHEMA || 'main',
       host: process.env.POSTGRES_TEST_HOST || 'localhost',
       port: process.env.POSTGRES_TEST_PORT || '5433',
       database: process.env.POSTGRES_TEST_DB || 'test-db',
