--- src/config/entities/index.ts
@@ -1,2 +1,0 @@
-export * from './orm.config';
-export * from './configuration';
