+++ src/datasources/accounts/accounts.datasource.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { PostgresDatabaseModule } from '@/datasources/db/postgres-database.module';
+import { PostgresDatabaseModule } from '@/datasources/db/v1/postgres-database.module';
 import { AccountsDatasource } from '@/datasources/accounts/accounts.datasource';
 import { IAccountsDatasource } from '@/domain/interfaces/accounts.datasource.interface';
