+++ src/datasources/db/v1/postgres-database.module.ts
@@ -1,12 +1,12 @@
 import postgres from 'postgres';
 import { Module } from '@nestjs/common';
-import { PostgresDatabaseShutdownHook } from '@/datasources/db/postgres-database.shutdown.hook';
+import { PostgresDatabaseShutdownHook } from '@/datasources/db/v1/postgres-database.shutdown.hook';
 import { IConfigurationService } from '@/config/configuration.service.interface';
-import { PostgresDatabaseMigrationHook } from '@/datasources/db/postgres-database.migration.hook';
+import { PostgresDatabaseMigrationHook } from '@/datasources/db/v1/postgres-database.migration.hook';
 import fs from 'fs';
-import { PostgresDatabaseMigrator } from '@/datasources/db/postgres-database.migrator';
-import { ICachedQueryResolver } from '@/datasources/db/cached-query-resolver.interface';
-import { CachedQueryResolver } from '@/datasources/db/cached-query-resolver';
+import { PostgresDatabaseMigrator } from '@/datasources/db/v1/postgres-database.migrator';
+import { ICachedQueryResolver } from '@/datasources/db/v1/cached-query-resolver.interface';
+import { CachedQueryResolver } from '@/datasources/db/v1/cached-query-resolver';
 
 function dbFactory(configurationService: IConfigurationService): postgres.Sql {
   const caPath = configurationService.get<string>('db.postgres.ssl.caPath');
