+++ src/datasources/notifications/notifications.datasource.module.ts
@@ -1,6 +1,6 @@
 import { AccountsDatasourceModule } from '@/datasources/accounts/accounts.datasource.module';
 import { NotificationsDatasource } from '@/datasources/notifications/notifications.datasource';
-import { PostgresDatabaseModule } from '@/datasources/db/postgres-database.module';
+import { PostgresDatabaseModule } from '@/datasources/db/v1/postgres-database.module';
 import { INotificationsDatasource } from '@/domain/interfaces/notifications.datasource.interface';
 import { Module } from '@nestjs/common';
