+++ src/domain/accounts/entities/account-data-type.entity.ts
@@ -1,4 +1,4 @@
-import { RowSchema } from '@/datasources/db/entities/row.entity';
+import { RowSchema } from '@/datasources/db/v1/entities/row.entity';
 import { z } from 'zod';
 
 export enum AccountDataTypeNames {
