+++ src/domain/accounts/entities/group.entity.ts
@@ -1,4 +1,4 @@
-import { RowSchema } from '@/datasources/db/entities/row.entity';
+import { RowSchema } from '@/datasources/db/v1/entities/row.entity';
 import type { z } from 'zod';
 
 export type Group = z.infer<typeof GroupSchema>;
