--- src/routes/notifications/v2/entities/index.ts
@@ -1,4 +1,0 @@
-export { NotificationType } from './notificationType.entity';
-export { NotificationDevice } from './notificationDevices.entity';
-export { NotificationSubscription } from './notificationSubscription.entity';
-export { NotificationSubscriptionNotificationType } from './notificationSubscriptionNotificationType.entity';
